+++ app/(payload)/admin/importMap.js
@@ -3,6 +3,7 @@ import { default as default_cfe8833dd4642d6046713de140337991 } from '@/component
 import { default as default_25d75f094cb6de9093023d79a509b859 } from '@/components/admin/UsersLoginMethodDisplay'
 import { default as default_3d5ce89c464db94aedb7076c07c0af9d } from '@/components/admin/HubspotSyncNowButton'
 import { default as default_ad497af92bb660355af21ff3a549e177 } from '@/components/admin/ApiKeyRowControls'
+import { default as default_ac52999068ac781bf4b426d2378d169a } from '@/components/admin/ResetContattiELogPanel'
 import { default as default_c147112e584c66ba0b5e4b5f7f822df0 } from '@/components/admin/CsvUploadNavLink'
 import { default as default_9c986bf00998f72758fbf6c76782cb67 } from '@/components/admin/GoogleAdminLoginButton'
 import { default as default_9490f073a9408ebc42a7406874d5bb3f } from '@/components/admin/LocalAdminLoginView'
@@ -16,6 +17,7 @@ export const importMap = {
   "@/components/admin/UsersLoginMethodDisplay#default": default_25d75f094cb6de9093023d79a509b859,
   "@/components/admin/HubspotSyncNowButton#default": default_3d5ce89c464db94aedb7076c07c0af9d,
   "@/components/admin/ApiKeyRowControls#default": default_ad497af92bb660355af21ff3a549e177,
+  "@/components/admin/ResetContattiELogPanel#default": default_ac52999068ac781bf4b426d2378d169a,
   "@/components/admin/CsvUploadNavLink#default": default_c147112e584c66ba0b5e4b5f7f822df0,
   "@/components/admin/GoogleAdminLoginButton#default": default_9c986bf00998f72758fbf6c76782cb67,
   "@/components/admin/LocalAdminLoginView#default": default_9490f073a9408ebc42a7406874d5bb3f,